% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2dimport.R
\name{c2dimport}
\alias{c2dimport}
\title{Import row data from an (unpacked) c2d file.}
\usage{
c2dimport(path = "path/to/your/example.xml")
}
\arguments{
\item{path}{Path to an XML file (i.e., an unpacked c2d file).}
}
\value{
dataframe containing the row data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is for basic testing.
The goal is to load directly from C2D files, not just unpacked XML.
}
\examples{
# data <- c2dimport("path/to/your/example.xml")

}
