% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2dimport.R
\name{c2dimport}
\alias{c2dimport}
\title{Import row data from an (unpacked) c2d file.}
\usage{
c2dimport(path = "path/to/your/example.xml")
}
\arguments{
\item{path}{Path to an XML file (i.e., an unpacked c2d file).}
}
\value{
dataframe containing the row data.
}
\description{
Import row data from an (unpacked) c2d file.
}
\examples{
# data <- c2dimport("path/to/your/example.xml")

}
