% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2dimport.R
\name{c2dimport}
\alias{c2dimport}
\title{Import row data from a Cyclus2 .c2d file.}
\usage{
c2dimport(path = "path/to/your/example.c2d")
}
\arguments{
\item{path}{Path to a C2D file (i.e., a gzipped XML file).}
}
\value{
dataframe containing the row data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is currently experimental.
The function loads the data Header and Rows from .c2d files.
}
\examples{
# rowdata <- c2dimport("path/to/your/example.c2d")

}
