% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2dimport.R
\name{c2dimport}
\alias{c2dimport}
\title{Import data from a Cyclus2 .c2d file.}
\usage{
c2dimport(path = "path/to/your/example.c2d")
}
\arguments{
\item{path}{Path to a C2D file (i.e., a gzipped XML file).}
}
\value{
A tibble containing the Athlete, Cycle, and Row data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Load the Athlete data, CycleData, and the Row data from a .c2d file
into a single tidyverse tibble.
This function is currently experimental, but hopefully already useful.
}
\examples{
# data <- c2dimport("path/to/your/example.c2d")

}
